Refactor registration form to standalone component
refactor register page to organized components like the login page
